--- Devantler.ContainerEngineProvisioner.slnx
@@ -1,16 +1,0 @@
-<Solution>
-  <Configurations>
-    <Platform Name="Any CPU" />
-    <Platform Name="x64" />
-    <Platform Name="x86" />
-  </Configurations>
-  <Folder Name="/src/">
-    <Project Path="src/Devantler.ContainerEngineProvisioner.Core/Devantler.ContainerEngineProvisioner.Core.csproj" />
-    <Project Path="src/Devantler.ContainerEngineProvisioner.Docker/Devantler.ContainerEngineProvisioner.Docker.csproj" />
-    <Project Path="src/Devantler.ContainerEngineProvisioner.Podman/Devantler.ContainerEngineProvisioner.Podman.csproj" />
-  </Folder>
-  <Folder Name="/tests/">
-    <Project Path="tests/Devantler.ContainerEngineProvisioner.Docker.Tests/Devantler.ContainerEngineProvisioner.Docker.Tests.csproj" />
-    <Project Path="tests/Devantler.ContainerEngineProvisioner.Podman.Tests/Devantler.ContainerEngineProvisioner.Podman.Tests.csproj" />
-  </Folder>
-</Solution>
