+++ src/DevantlerTech.ContainerEngineProvisioner.Podman/DevantlerTech.ContainerEngineProvisioner.Podman.csproj
@@ -12,7 +12,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Devantler.ContainerEngineProvisioner.Docker\Devantler.ContainerEngineProvisioner.Docker.csproj" />
+    <ProjectReference Include="..\..\src\DevantlerTech.ContainerEngineProvisioner.Docker\DevantlerTech.ContainerEngineProvisioner.Docker.csproj" />
   </ItemGroup>
 
 </Project>
