+++ tests/DevantlerTech.ContainerEngineProvisioner.Podman.Tests/PodmanProvisionerTests/CheckReadyAsyncTests.cs
@@ -1,4 +1,4 @@
-namespace Devantler.ContainerEngineProvisioner.Podman.Tests.PodmanProvisionerTests;
+namespace DevantlerTech.ContainerEngineProvisioner.Podman.Tests.PodmanProvisionerTests;
 
 /// <summary>
 /// Unit tests for <see cref="PodmanProvisioner.CheckReadyAsync(CancellationToken)"/>.
