+++ tests/DevantlerTech.ContainerEngineProvisioner.Podman.Tests/PodmanProvisionerTests/CreateDirectoryInContainerAsyncTests.cs
@@ -1,6 +1,6 @@
 using Docker.DotNet.Models;
 
-namespace Devantler.ContainerEngineProvisioner.Podman.Tests.PodmanProvisionerTests;
+namespace DevantlerTech.ContainerEngineProvisioner.Podman.Tests.PodmanProvisionerTests;
 
 /// <summary>
 /// Tests for <see cref="PodmanProvisioner"/>.
